got all the documentation up to standar, fixed some clang tidy errors and a lot of typos

+++ SFMLConversion/Source/entities/SFMLBonusCar.cpp
@@ -1,6 +1,7 @@
 /**
  *@file
  * implementation of the sfmlBonusCar class
+ * @author Thibaut Van Goethem
  */
 
 #include "../../Include/Entities/SFMLBonusCar.h"
@@ -19,12 +20,12 @@ namespace roadfighterSFML {
                                double vertSpeed)
             : SFMLEntitySprite("./resources/sprites/special_car.png", window),
               BonusCar(m_loc1, m_loc2, vertSpeed) {
-        std::tuple<int, int> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
-        std::tuple<int, int> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
+        std::tuple<double, double> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
+        std::tuple<double, double> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
         setOrigin(SFMLEntitySprite::getGlobalBounds().width / 2, SFMLEntitySprite::getGlobalBounds().height / 2);
         SFMLEntitySprite::scale(sf::Vector2f(
-                ((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
-                ((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
+                static_cast<float>((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
+                static_cast<float>((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
     }
 
     /**
@@ -33,7 +34,7 @@ namespace roadfighterSFML {
      * @exception none
      */
     void SFMLBonusCar::draw() {
-        std::tuple<int, int> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
+        std::tuple<double, double> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
         setSpriteLocation(std::get<0>(newloc) + SFMLEntitySprite::getGlobalBounds().width / 2,
                           std::get<1>(newloc) + SFMLEntitySprite::getGlobalBounds().height / 2);
         if (getVerticalSpeed() < 0) {
