+++ SFMLConversion/Source/entities/SFMLEnd.cpp
@@ -1,6 +1,7 @@
 /**
  *@file
  * implementation of the SFMLEnd class
+ * @author Thibaut Van Goethem
  */
 #include "../../Include/Entities/SFMLEnd.h"
 namespace roadfighterSFML {
@@ -8,7 +9,7 @@ namespace roadfighterSFML {
     /**
      * constructor for the sfmlend class
      * @param m_loc1 first location of the end
-     * @param m_loc2 second laoction of the object
+     * @param m_loc2 second location of the object
      * @param window the sfmlrenderwindow that is used to draw on
      * @return none
      * @exception none
@@ -18,11 +19,11 @@ namespace roadfighterSFML {
                                                                         SFMLEntitySprite(
                                                                                 "./resources/sprites/Finishline.png",
                                                                                 window) {
-        std::tuple<int, int> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
-        std::tuple<int, int> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
+        std::tuple<double, double> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
+        std::tuple<double, double> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
         SFMLEntitySprite::scale(sf::Vector2f(
-                ((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
-                ((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
+                static_cast<float>((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
+                static_cast<float>((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
     }
     /**
       * draw function that is overriden from roadfighter::entity
@@ -30,7 +31,7 @@ namespace roadfighterSFML {
       * @exception none
       */
     void SFMLEnd::draw() {
-        std::tuple<int, int> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
+        std::tuple<double, double> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
         setSpriteLocation(std::get<0>(newloc), std::get<1>(newloc));
         getWindow()->draw(*this);
     }
