+++ SFMLConversion/Source/entities/SFMLPassingCar.cpp
@@ -1,6 +1,7 @@
 /**
  *@file
  * implementation of the SFMLPassingCar class
+ * @author Thibaut Van Goethem
  */
 
 #include "../../Include/Entities/SFMLPassingCar.h"
@@ -18,12 +19,12 @@ namespace roadfighterSFML {
                                    double vertSpeed) : SFMLEntitySprite(
             "./resources/sprites/passing_car.png", window),
                                                        PassingCar(m_loc1, m_loc2, vertSpeed) {
-        std::tuple<int, int> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
-        std::tuple<int, int> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
+        std::tuple<double, double> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
+        std::tuple<double, double> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
         setOrigin(SFMLEntitySprite::getGlobalBounds().width / 2, SFMLEntitySprite::getGlobalBounds().height / 2);
         SFMLEntitySprite::scale(sf::Vector2f(
-                ((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
-                ((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
+                static_cast<float>((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
+                static_cast<float>((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
     }
 
     /**
@@ -32,7 +33,7 @@ namespace roadfighterSFML {
       * @exception none
       */
     void SFMLPassingCar::draw() {
-        std::tuple<int, int> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
+        std::tuple<double, double> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
         setSpriteLocation(std::get<0>(newloc) + SFMLEntitySprite::getGlobalBounds().width / 2,
                           std::get<1>(newloc) + SFMLEntitySprite::getGlobalBounds().height / 2);
         if (getVerticalSpeed() < 0) {
