+++ SFMLConversion/Source/entities/SFMLPlayerCar.cpp
@@ -1,6 +1,7 @@
 /**
  *@file
  * implementation of the SFMLPlayerCar class
+ * @author Thibaut Van Goethem
  */
 
 #include "../../Include/Entities/SFMLPlayerCar.h"
@@ -15,7 +16,7 @@ namespace roadfighterSFML {
      * @param m_horAccel the horizontal acceleration of the car
      * @param m_fuel the starting fuel
      * @param m_moveController the controller that is going to be used to get input from
-     * @param transporter the entity tranporter that the car will use to make bullets and get them to the world
+     * @param transporter the entity transporter that the car will use to make bullets and get them to the world
      * @param factory the entity factory that the car will use to create the bullets
      * @param window the renderwindow that will be used to draw on
      * @return none
@@ -31,12 +32,12 @@ namespace roadfighterSFML {
                                                                                     SFMLEntitySprite(
                                                                                             "./resources/sprites/player_car.png",
                                                                                             window) {
-        std::tuple<int, int> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
-        std::tuple<int, int> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
+        std::tuple<double, double> sfmlpos1 = Transformation::getInstance().locationTransformation(getLoc1());
+        std::tuple<double, double> sfmlpos2 = Transformation::getInstance().locationTransformation(getLoc2());
         setOrigin(SFMLEntitySprite::getGlobalBounds().width / 2, SFMLEntitySprite::getGlobalBounds().height / 2);
         SFMLEntitySprite::scale(sf::Vector2f(
-                ((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
-                ((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
+                static_cast<float>((std::get<0>(sfmlpos2) - std::get<0>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().width),
+                static_cast<float>((std::get<1>(sfmlpos2) - std::get<1>(sfmlpos1)) / SFMLEntitySprite::getGlobalBounds().height)));
 
     }
 
@@ -46,7 +47,7 @@ namespace roadfighterSFML {
     * @exception none
     */
     void SFMLPlayerCar::draw() {
-        std::tuple<int, int> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
+        std::tuple<double, double> newloc = Transformation::getInstance().locationTransformation(this->getLoc1());
         setSpriteLocation(std::get<0>(newloc) + SFMLEntitySprite::getGlobalBounds().width / 2,
                           std::get<1>(newloc) + SFMLEntitySprite::getGlobalBounds().height / 2);
         if (getHorizontalSpeed() < 0) {
